rebase master + rename contentful reminders.data provider -> reminder.data provider
